fix(android): fix Fabric cannot be enabled on 0.72
